Add AnnotateModels::SchemaInfo
