+++ spec/annotate/annotate_models_spec.rb
@@ -63,19 +63,6 @@ describe AnnotateModels do
     double('Column', stubs)
   end
 
-  describe '#quote' do
-    it 'converts values to the default column value as a string' do
-      expect(AnnotateModels.send(:quote, nil)).to eql('NULL')
-      expect(AnnotateModels.send(:quote, true)).to eql('TRUE')
-      expect(AnnotateModels.send(:quote, false)).to eql('FALSE')
-      expect(AnnotateModels.send(:quote, 25)).to eql('25')
-      expect(AnnotateModels.send(:quote, 25.6)).to eql('25.6')
-      expect(AnnotateModels.send(:quote, 1e-20)).to eql('1.0e-20')
-      expect(AnnotateModels.send(:quote, BigDecimal.new('1.2'))).to eql('1.2')
-      expect(AnnotateModels.send(:quote, [BigDecimal.new('1.2')])).to eql(['1.2'])
-    end
-  end
-
   describe '#parse_options' do
     let(:options) do
       {
@@ -1769,6 +1756,21 @@ end
 
     it 'skips attempt to annotate if no table exists for model' do
       expect(subject).to be_nil
+    end
+  end
+
+  describe 'SchemaInfo' do
+    describe '.#quote' do
+      it 'converts values to the default column value as a string' do
+        expect(AnnotateModels::SchemaInfo.send(:quote, nil)).to eql('NULL')
+        expect(AnnotateModels::SchemaInfo.send(:quote, true)).to eql('TRUE')
+        expect(AnnotateModels::SchemaInfo.send(:quote, false)).to eql('FALSE')
+        expect(AnnotateModels::SchemaInfo.send(:quote, 25)).to eql('25')
+        expect(AnnotateModels::SchemaInfo.send(:quote, 25.6)).to eql('25.6')
+        expect(AnnotateModels::SchemaInfo.send(:quote, 1e-20)).to eql('1.0e-20')
+        expect(AnnotateModels::SchemaInfo.send(:quote, BigDecimal.new('1.2'))).to eql('1.2')
+        expect(AnnotateModels::SchemaInfo.send(:quote, [BigDecimal.new('1.2')])).to eql(['1.2'])
+      end
     end
   end
 end
